+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [🚀] `Capture.capture`增加指定的`pixelRatio`参数;
 - [🚀] `IterableFutureX`增加`thenTake`和`thenSkip`扩展;
 - [🚀] 集成`auto_size_text`;
+- [🚀] 增加`no_transition_page_builder`;
 
 ## 0.51.0
 - [🚀] [break change] 移除`hive`;